Creating a new SceneQueItem interface that wraps any QueItem that has an OBS scene.
Right now the core relies directly on the OBS Plugin to do some basic things.  This allows other plugins to extend an OBS-scene type object.

Note: probably a better solve is to remove all the nextQueueScene/previewScene logic from the core entirely and move to a plugin but the plugins still need some way to know that an item is an OBS scene